# backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_api_rest_v1.AuthResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/internal_api_rest_v1.UserResponse'
    type: object
  internal_api_rest_v1.ChangePasswordResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
...
  internal_api_rest_v1.GitHubLoginRequest:
    properties:
      code:
        type: string
      state:
        type: string
    required:
    - code
    - state
    type: object
...
  internal_api_rest_v1.MessageResponse:
    properties:
      message:
        type: string
    type: object
  internal_api_rest_v1.OAuthURLResponse:
    properties:
      state:
        type: string
      url:
        type: string
    type: object
...
  internal_api_rest_v1.PasswordChangeRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - new_password
    type: object
...
  internal_api_rest_v1.PostKeyResponse:
    properties:
      created_at:
        type: string
      post_key:
        type: string
    type: object
...
  internal_api_rest_v1.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_api_rest_v1.RefreshTokenResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
  internal_api_rest_v1.RotatePostKeyResponse:
    properties:
      post_key:
        type: string
    type: object
  internal_api_rest_v1.SessionsResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/markpost_internal_domain_user.RefreshToken'
        type: array
    type: object
...
  internal_api_rest_v1.UserResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      name:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  internal_api_rest_v1.UsernameLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  markpost_internal_apierr.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items:
          $ref: '#/definitions/markpost_internal_apierr.FieldError'
        type: array
      message:
        type: string
    type: object
  markpost_internal_apierr.FieldError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
...
  markpost_internal_domain_user.RefreshToken:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      revoked:
        type: boolean
      user_id:
        type: integer
    type: object
host: localhost:7330
info:
  contact: {}
  description: Markpost backend API for post management and delivery.
  title: Markpost API
  version: "1.0"
paths:
...
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_rest_v1.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.ChangePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change the current user's password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_rest_v1.UsernameLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      summary: Login with username and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout the current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_rest_v1.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.RefreshTokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      summary: Refresh authentication token
      tags:
      - auth
  /api/v1/auth/sessions:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke all of the current user's sessions except the present one (I.12)
      tags:
      - auth
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.SessionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the current user's sessions (I.12)
      tags:
      - auth
  /api/v1/auth/sessions/{token_id}:
    delete:
      parameters:
      - description: Session (refresh token) ID
        in: path
        name: token_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke one of the current user's sessions (I.12)
      tags:
      - auth
...
  /api/v1/oauth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: GitHub OAuth code and state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_rest_v1.GitHubLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      summary: Login with GitHub OAuth code
      tags:
      - oauth
  /api/v1/oauth/url:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.OAuthURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      summary: Get GitHub OAuth authorization URL
      tags:
      - oauth
  /api/v1/post-key/rotate:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.RotatePostKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rotate the current user's post key (C2.5)
      tags:
      - auth
  /api/v1/post_key:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_rest_v1.PostKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/markpost_internal_apierr.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the current user's post key
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: '"Bearer <access_token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"